Create merge_binary_2.cpp
This code compiles all of the text data files from the oscilloscope and puts them into a compact binary file. The code can be run with arguments, which are as follows:
1st argument: folder to read from
2nd argument: folder to write to
3rd argument (not recommended to be used): name stored within binary file
As an example, to merge and create the files within the current directory, assuming we named our executable "merge," we would run the following command on the CERNBOX terminal:
bash-4.2$ ./merge $(pwd)

The file output is called 'binary_dat.' This can easily be changed if wanted.