Fix to merge_num.sh
Updated to now rename files in each subfile so that numbering goes from 0 to 5000. This allows it to work with merge_binary_3.cpp. The alternative would have been to modify merge_binary_3.cpp to accept a second argument, the number to start on, but I consulted with Robert and he preferred this option.